feat : Refactor books methods to respect SOLID
