+++ library/books/book.rb
@@ -1,6 +1,4 @@
-require './library/rental'
-
-# Book class
+require_relative 'rental'
 
 class Book
   attr_accessor :title, :author, :rentals
