+++ library/main.rb
@@ -8,20 +8,19 @@ require './library/rental'
 
 class Library
   def initialize
-    @books = []
     @people = []
     @rentals = []
   end
 
-  def list_all_books
-    if @books.empty?
-      puts 'Books list is empty, click on number 4 to create a book'
-      return
-    end
-    @books.each do |book|
-      print "Title: #{book.title.capitalize}, Author: #{book.author.capitalize}\n"
-    end
-  end
+  #    def list_all_books
+  #     if @books.empty?
+  #       puts 'Books list is empty, click on number 4 to create a book'
+  #       return
+  #     end
+  #     @books.each do |book|
+  #       print "Title: #{book.title.capitalize}, Author: #{book.author.capitalize}\n"
+  #     end
+  #   end
 
   def list_all_people
     if @people.empty?
@@ -71,17 +70,17 @@ class Library
     end
   end
 
-  def create_book
-    print 'Title: '
-    title = gets.chomp.capitalize
-    print 'Author: '
-    author = gets.chomp.capitalize
-
-    book = Book.new(title, author)
-    @books << book
-
-    puts "Book created successfully\n"
-  end
+  #   def create_book
+  #     print 'Title: '
+  #     title = gets.chomp.capitalize
+  #     print 'Author: '
+  #     author = gets.chomp.capitalize
+  #
+  #     book = Book.new(title, author)
+  #     @books << book
+  #
+  #     puts "Book created successfully\n"
+  #   end
 
   def create_rental
     if @people.empty? && @books.empty?
