+++ library/people/person.rb
