+++ library/people/teacher.rb
