+++ library/rentals/rental.rb
